대체 뭐가 문제였을까

+++ HyruleEngine/GameObject.h
@@ -21,7 +21,7 @@ namespace Hyrule
 		std::wstring tag;								// 태그
 		std::wstring name;								// 이름
 
-		std::map<unsigned int, Component*> components;	// 컴포넌트 리스트
+		std::map<size_t, Component*> components;	// 컴포넌트 리스트
 		Transform* transform;							// 트랜스폼
 		
 		// 유니티에서는 게임 오브젝트가 씬에 접근할 수 있었는데...
